requirements.txt
argon2-cffi==21.3.0
email_validator==2.2.0
fastapi==0.115.14
openai==1.93.0
password_strength==0.0.3.post2
pydantic==2.11.7
pymongo==4.13.2
python-socketio==5.13.0
python_jose==3.5.0
validators==0.35.0
